Refactored the include directory out
