+++ include/pocketbase/services/baseauthstore.h
