+++ include/pocketbase/services/sseclient.h
